requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
kafka-python>=2.3.1
psycopg[binary]>=3.0.0